requirements.txt
Pillow==3.4.2
lmdb==0.92
torch==0.1.8.post1
torchvision==0.1.7